refactor: list, mapper로 공통 상수 관리
