--- src/constants/PeerTypeCard.ts
@@ -1,18 +1,0 @@
-export const PeerTypeCard = {
-  D: {
-    caption: '본투비 리더형',
-    title: '성취하는 행동 대장',
-  },
-  I: {
-    caption: '본투비 인싸',
-    title: '긍정적인 분위기 메이커',
-  },
-  S: {
-    caption: '본투비 평화주의자',
-    title: '화합을 도모하는 조력자',
-  },
-  C: {
-    caption: '본투비 분석가',
-    title: '신중한 설계자',
-  },
-};
