--- src/constants/noticeType.ts
@@ -1,17 +1,0 @@
-export enum NoticeType {
-  INVITE_TO_PROJECT = 'INVITE_TO_PROJECT',
-  REQUEST_JOIN_PROJECT = 'REQUEST_JOIN_PROJECT',
-  ACCEPT_PROJECT_INVITATION = 'ACCEPT_PROJECT_INVITATION',
-  DECLINE_PROJECT_INVITATION = 'DECLINE_PROJECT_INVITATION',
-  ACCEPT_PROJECT_JOIN_REQUEST = 'ACCEPT_PROJECT_JOIN_REQUEST',
-  DECLINE_PROJECT_JOIN_REQUEST = 'DECLINE_PROJECT_JOIN_REQUEST',
-}
-
-export const NOTICE_MAPPER: Record<NoticeType, string> = {
-  INVITE_TO_PROJECT: '프로젝트 초대',
-  REQUEST_JOIN_PROJECT: '프로젝트 참여 신청',
-  ACCEPT_PROJECT_INVITATION: '프로젝트 초대 수락',
-  DECLINE_PROJECT_INVITATION: '프로젝트 초대 거절',
-  ACCEPT_PROJECT_JOIN_REQUEST: '프로젝트 참여 요청 수락',
-  DECLINE_PROJECT_JOIN_REQUEST: '프로젝트 참여 요청 거절',
-};
