--- src/constants/onboard.ts
@@ -1,24 +1,0 @@
-import Onboarding1 from '@assets/onboarding/Onboarding1.png';
-import Onboarding2 from '@assets/onboarding/Onboarding2.png';
-import Onboarding3 from '@assets/onboarding/Onboarding3.png';
-import Onboarding4 from '@assets/onboarding/Onboarding4.png';
-
-import Peer1 from '@assets/review/peer/peer1.png';
-import Peer2 from '@assets/review/peer/peer2.png';
-import Peer3 from '@assets/review/peer/peer3.png';
-
-export const ON_BOARDING_HEADER_TEXT = [
-  '당신의 동료가 피어나는 곳',
-  '동료가 직접 응답해서 \n 더 정확하고 객관적인 내 프로필',
-  '나는 어떤 동료와 잘 맞을까? \n 동료와 나의 직관적인 비교!',
-  '잘 맞는 동료와 성공적인 프로젝트 \n 피어나에서 시작하세요!',
-];
-
-export const ON_BOARDING_IMAGE_LIST = [
-  Onboarding1,
-  Onboarding2,
-  Onboarding3,
-  Onboarding4,
-];
-
-export const PEER_REVIEW_IMAGE_LIST = [Peer1, Peer2, Peer3];
